+++ mostlyai/sdk/_local/synthetic_datasets.py
@@ -25,7 +25,7 @@ from mostlyai.sdk._local.execution.plan import (
     has_tabular_model,
     has_language_model,
     FINALIZE_GENERATION_TASK_STEPS,
-    MODEL_TYPE_STEPS_MAP,
+    get_model_type_generation_steps_map,
 )
 from mostlyai.sdk.client._base_utils import convert_to_df
 from mostlyai.sdk.domain import (
@@ -111,11 +111,14 @@ def create_synthetic_dataset(
     # copy ModelQA reports into synthetic dataset directory
     source_reports_dir = generator_dir / "ModelQAReports"
     dest_reports_dir = synthetic_dataset_dir / "ModelQAReports"
-    shutil.copytree(source_reports_dir, dest_reports_dir)
+    if source_reports_dir.exists():
+        shutil.copytree(source_reports_dir, dest_reports_dir)
 
     # create job progress
     progress_steps: list[ProgressStep] = []
     for table in generator.tables:
+        sd_table = next(t for t in config.tables if t.name == table.name)
+        steps_map = get_model_type_generation_steps_map(sd_table.configuration.enable_data_report)
         model_types = [
             model_type
             for model_type, check in [
@@ -125,7 +128,7 @@ def create_synthetic_dataset(
             if check
         ]
         for model_type in model_types:
-            for step in MODEL_TYPE_STEPS_MAP[model_type]:
+            for step in steps_map[model_type]:
                 progress_steps.append(
                     ProgressStep(
                         task_type=TaskType.generate,
